+++ src/onegov.core/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds request reference to form.meta.
+  [href]
+
 0.74.5 (2019-01-03)
 ~~~~~~~~~~~~~~~~~~~